requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-celery-results==2.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-grappelli==2.15.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.21.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.2
prompt-toolkit==3.0.18
pydantic==1.8.2
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
